Fix GdClean and add Trash Up

+++ bot/helper/mirror_utils/upload_utils/gdriveTools.py
@@ -181,7 +181,7 @@ class GoogleDriveHelper:
             msg = str(err)
         return msg
         
-    def driveclean(self, drive_id):
+    def driveclean(self, drive_id: str, trash: bool):
         msg = ''
         query = f"'{drive_id}' in parents and trashed = false"
         page_token = None
@@ -192,7 +192,10 @@ class GoogleDriveHelper:
                 for file in files:
                     self.__total_files += 1
                     self.__total_bytes += int(file.get('size', 0))
-                    self.__service.files().delete(fileId=file['id'], supportsAllDrives=True).execute()
+                    if trash:
+                        self.__service.files().update(fileId=file['id'], body={"trashed": True}).execute()
+                    else:
+                        self.__service.files().delete(fileId=file['id'], supportsAllDrives=True).execute()
                 page_token = drive_query.get('nextPageToken', None)
                 if page_token is None:
                     msg = f"⌬ <b><i>Successfully Cleaned Folder/Drive :</i></b> \n\n<b>Total Files:</b> <code>{self.__total_files}</code>\n<b>Total Size:</b> <code>{get_readable_file_size(self.__total_bytes)}</code>"
